+++ internal/commands/account/delete_test.go
@@ -1,0 +1,52 @@
+package account
+
+import (
+	"testing"
+
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands"
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/config"
+	smock "github.com/UpCloudLtd/upcloud-cli/v3/internal/mock"
+
+	"github.com/UpCloudLtd/upcloud-go-api/v6/upcloud"
+	"github.com/UpCloudLtd/upcloud-go-api/v6/upcloud/request"
+	"github.com/gemalto/flume"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestDeleteCommand(t *testing.T) {
+	targetMethod := "DeleteSubaccount"
+
+	account := upcloud.Account{
+		UserName: "my-sub-account",
+	}
+
+	for _, test := range []struct {
+		name  string
+		arg   string
+		error string
+		req   request.DeleteSubaccountRequest
+	}{
+		{
+			name: "delete with username",
+			arg:  account.UserName,
+			req:  request.DeleteSubaccountRequest{Username: account.UserName},
+		},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			mService := smock.Service{}
+			mService.On(targetMethod, &test.req).Return(nil)
+
+			conf := config.New()
+			c := commands.BuildCommand(DeleteCommand(), nil, conf)
+
+			_, err := c.(commands.MultipleArgumentCommand).Execute(commands.NewExecutor(conf, &mService, flume.New("test")), test.arg)
+
+			if test.error != "" {
+				assert.EqualError(t, err, test.error)
+			} else {
+				assert.NoError(t, err)
+				mService.AssertNumberOfCalls(t, targetMethod, 1)
+			}
+		})
+	}
+}