+++ internal/completion/account_test.go
@@ -1,0 +1,50 @@
+package completion_test
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/completion"
+	smock "github.com/UpCloudLtd/upcloud-cli/v3/internal/mock"
+	"github.com/UpCloudLtd/upcloud-go-api/v6/upcloud"
+	"github.com/spf13/cobra"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+)
+
+var mockAccounts = upcloud.AccountList{
+	{Username: "mock1"},
+	{Username: "mock2"},
+	{Username: "bock1"},
+	{Username: "bock2"},
+	{Username: "dock1"},
+}
+
+func TestAccount_CompleteArgument(t *testing.T) {
+	for _, test := range []struct {
+		name              string
+		complete          string
+		expectedMatches   []string
+		expectedDirective cobra.ShellCompDirective
+	}{
+		{name: "basic username", complete: "dock", expectedMatches: []string{"dock1"}, expectedDirective: cobra.ShellCompDirectiveNoFileComp},
+		{name: "multiple usernames", complete: "mock", expectedMatches: []string{"mock1", "mock2"}, expectedDirective: cobra.ShellCompDirectiveNoFileComp},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			mService := new(smock.Service)
+			mService.On("GetAccountList", mock.Anything).Return(mockAccounts, nil)
+			accounts, directive := completion.Account{}.CompleteArgument(context.TODO(), mService, test.complete)
+			assert.Equal(t, test.expectedMatches, accounts)
+			assert.Equal(t, test.expectedDirective, directive)
+		})
+	}
+}
+
+func TestAccount_CompleteArgumentServiceFail(t *testing.T) {
+	mService := new(smock.Service)
+	mService.On("GetAccountList", mock.Anything).Return(nil, fmt.Errorf("MOCKFAIL"))
+	ips, directive := completion.Account{}.CompleteArgument(context.TODO(), mService, "127")
+	assert.Nil(t, ips)
+	assert.Equal(t, cobra.ShellCompDirectiveDefault, directive)
+}